Fix loading position of song repo entries
